+++ ebpub/ebpub/db/tests/test_views.py
@@ -121,6 +121,7 @@ class LocationDetailTestCase(BaseTestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.context['newsitem_list']), 1)
         self.assertEqual(response.context['newsitem_list'][0].title, 'crime title 1')
+        # We don't get 'crime title 2' in this query because it's in the future.
 
     def test_location_overview(self):
         url = urlresolvers.reverse('ebpub-location-overview',
@@ -261,16 +262,6 @@ class TestSchemaFilterView(BaseTestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 302)
         self.assertEqual(response['location'], 'http://testserver/whee/')
-
-    def test_filter__charting_disallowed_redirect(self):
-        # could make a different fixture, but, meh.
-        crime = models.Schema.objects.get(slug='crime')
-        crime.allow_charting = False
-        crime.save()
-        url = filter_reverse('crime', [('by-foo', 'bar')])
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 301)
-        self.assertEqual(response['location'], 'http://testserver/crime/')
 
     def test_filter__bad_params(self):
         url = filter_reverse('crime', [('by-foo', 'bar')])