+++ ebpub/ebpub/richmaps/views.py
@@ -16,35 +16,33 @@
 #   along with ebpub.  If not, see <http://www.gnu.org/licenses/>.
 #
 
+from django import template
 from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse, HttpResponseRedirect
-from django import template
 from django.shortcuts import get_object_or_404
 from django.template.loader import get_template, select_template
-from django.utils.cache import patch_response_headers
-from django.utils.datastructures import SortedDict
 from django.utils import simplejson
-from ebpub.utils.view_utils import eb_render
-from ebpub.db.models import NewsItem, Schema, SchemaField
-from ebpub.openblockapi.views import _copy_nomulti, JSON_CONTENT_TYPE
-from ebpub.openblockapi.itemquery import build_item_query
-from ebpub.db.schemafilters import FilterError
+from django.utils.cache import patch_response_headers
+from ebpub.db.models import NewsItem
 from ebpub.db.schemafilters import FilterChain
-from ebpub.db.schemafilters import BadAddressException
-from ebpub.db.schemafilters import BadDateException
 from ebpub.db.views import _get_filter_schemafields
+from ebpub.openblockapi.itemquery import build_item_query
+from ebpub.openblockapi.views import JSON_CONTENT_TYPE
 from ebpub.streets.models import Place, PlaceType
+from ebpub.utils.view_utils import eb_render
 from ebpub.utils.view_utils import get_schema_manager
 import datetime
 import logging
 import re
 
-
 logger = logging.getLogger('ebpub.richmaps.views')
 
 def bigmap_filter(request, slug):
-
+    """
+    Big map with just one Schema (identified by ``slug``) enabled by
+    default.
+    """
     s = get_object_or_404(get_schema_manager(request), slug=slug, is_special_report=False)
     if not s.allow_charting:
         return HttpResponse(status=404)
@@ -62,78 +60,87 @@ def bigmap_filter(request, slug):
 
     config = _decode_map_permalink(request, show_default_layers=False, filters=filterchain)
 
+
+    # TODO: This can leave in permalink params eg. 'i', even if there
+    # is also 'ids', because it doesn't recognize those as being the
+    # same.
     new_url = filterchain.make_url(base_url=reverse('bigmap_filter', args=(slug,)))
     if new_url != request.get_full_path():
-        return HttpResponseRedirect(new_url)    
+        return HttpResponseRedirect(new_url)
 
-    
-    # add in the filter layer
-    base_url = reverse('ebpub-schema-filter-geojson', args=(slug,))
-    layer_url = filterchain.make_url(base_url=base_url)
-    custom_layer = {
-        'url': layer_url,
-        'params': {},
-        'title': "Custom Filter",
-        'visible': True
-    }
-    config['layers'].append(custom_layer)
-
-
-
-    if config['is_widget']: 
+    if config['is_widget']:
         return eb_render(request, 'richmaps/embed_bigmap.html', {
             'map_config': simplejson.dumps(config, indent=2)
         })
-    else:         
+    else:
         return eb_render(request, 'richmaps/bigmap.html', {
             'map_config': simplejson.dumps(config, indent=2)
         })
 
 
 def bigmap(request):
-    config = _decode_map_permalink(request)
+    '''
+    Big map with all Schemas enabled by default.
+    '''
+    filterchain = FilterChain(request=request)
+    config = _decode_map_permalink(request, filters=filterchain)
 
     if config['is_widget']: 
         return eb_render(request, 'richmaps/embed_bigmap.html', {
             'map_config': simplejson.dumps(config, indent=2)
         })
-    else:         
+    else:
         return eb_render(request, 'richmaps/bigmap.html', {
             'map_config': simplejson.dumps(config, indent=2)
         })
 
 def _decode_map_permalink(request, show_default_layers=True, filters=None):
     """
-    request parameters: 
-    c - map center 
-    z - map zoom 
-    l - layer info
-    p - popup center
-    f - popup feature
+    Permalinks for the big map, with more compact query parameters.
+
+    Returns a map_config dictionary.
+
+    Accepted parameters:
+
+    c - map center, separated by underscore, eg. c=-92.28283_38.95658
+
+    z - map zoom, eg. z=12
+
+    l - layers to display on load, comma- or dash-separated,
+        eg. l=p13,t32,p1 or eg. l=p12345-t7-t9,
+        where p => place layer
+        and t => schema ("type") layer
+
+    i - items to load specificially by id, comma- or dash-separated,
+        eg. i=t1234-t456
+
+    p - popup center, with underscore, eg. p=-92.3438_38.9658
+    f - popup feature, eg. f=t1234 or f=p1234
+        where p = a place and t = a news item
+
     start_date - start date (inclusive) %m/%d/%Y
     end_date - end date (inclusive) %m/%d/%Y
-    d - duration in days (overridden by end date)
+    d - duration in days (overridden by end date), eg. d=7
 
-    x - show as 'widget' 
-    v- limits what map controls are displayed (widget only)
-    w - width of map (widget only)
-    h - height of map (widget only)
+    x - show as 'widget', just the map and nothign around it.
+        Takes no value, eg. x
+    w - width of map (widget only), in pixels
+    h - height of map (widget only), in pixels
+    v - limits what map controls are displayed (widget only).
+        By default, widget-stype map shows none of these.
+        Possible values, joined with no separator:
+        l - layer switcher
+        h - list of headlines next to map
+        p - permalink
+        eg. to turn them all on: v=lhp
+
+
     """
-    
     params = request.GET
-    
-    #
-    # enabled item, place layers "l" parameter
-    #
-    # l=p13,t32,p1 ...
-    #
-    # p => place layer
-    # t => schema layer 
-    # remaining portion is dbid of place or schema
-    #
     schemas = set()
     place_types = set()
     lids = params.get("l", None)
+    show_custom_layer = False
     if lids is not None: 
         no_layers_specified = False
         try:
@@ -145,9 +152,11 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
                     place_types.add(layer_id)
                 elif layer_type == 't': 
                     schemas.add(layer_id)
+                elif layer_type == 'c':
+                    show_custom_layer = True
         except: 
             pass
-    else: 
+    else:
         no_layers_specified = True
 
     # map center
@@ -188,39 +197,48 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
         except: 
             popup_center = None
             popup_feature = None
-    
 
     # start and end date range
-    startdate = params.get('start_date')
-    if startdate is not None:
-        try:
-            startdate = datetime.datetime.strptime(startdate, '%m/%d/%Y').date()
-        except:
-            startdate = None
-
-    enddate = params.get('end_date')
-    if enddate is not None: 
-        try:
-            enddate = datetime.datetime.strptime(enddate, '%m/%d/%Y').date()
-        except:
-            enddate = None
-
-    if startdate is None and enddate is None and filters:
-        date_filter = filters.get('date') or filters.get('pubdate')
-        if date_filter:
-            startdate = date_filter.start_date
-            enddate = date_filter.end_date
-
-
     default_interval = datetime.timedelta(days=7)
     duration = params.get('d')
     if duration is not None:
         try:
             duration = datetime.timedelta(days=int(duration))
-        except:
+        except (TypeError, ValueError):
             duration = default_interval
-    else: 
+    else:
         duration = default_interval
+    default_enddate = datetime.date.today()
+    default_startdate = default_enddate - duration
+
+    startdate = params.get('start_date')
+    if startdate is not None:
+        for format in ('%m/%d/%Y', '%Y-%m-%d'):
+            try:
+                startdate = datetime.datetime.strptime(startdate, format).date()
+                break
+            except ValueError:
+                pass
+        if isinstance(startdate, basestring):
+            startdate = None
+
+    enddate = params.get('end_date')
+    if enddate is not None:
+        for format in ('%m/%d/%Y', '%Y-%m-%d'):
+            try:
+                enddate = datetime.datetime.strptime(enddate, format).date()
+                break
+            except ValueError:
+                pass
+        if isinstance(enddate, basestring):
+            enddate = None
+
+    # The filters argument can override startdate & enddate.
+    if startdate is None and enddate is None and filters:
+        date_filter = filters.get('date') or filters.get('pubdate')
+        if date_filter:
+            startdate = date_filter.start_date
+            enddate = date_filter.end_date
 
     if startdate is None and enddate is None:
         enddate = datetime.date.today()
@@ -237,7 +255,15 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
     if filters and filters.get('date') is None: 
         filters.add('date', startdate, enddate)
 
+    api_startdate = startdate.strftime("%Y-%m-%d")  
+    api_enddate = (enddate + datetime.timedelta(days=1)).strftime("%Y-%m-%d")
+
     layers = []
+
+    if (startdate != default_startdate) or (enddate != default_enddate):
+        show_custom_layer = True
+
+    # All available place layers.
     for place_type in PlaceType.objects.filter(is_mappable=True).all():
         layers.append({
             'id': 'p%d' % place_type.id,
@@ -249,10 +275,17 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
             'visible': place_type.id in place_types # off by default
         })
 
-    api_startdate = startdate.strftime("%Y-%m-%d")  
-    api_enddate = (enddate + datetime.timedelta(days=1)).strftime("%Y-%m-%d")
-
+    # All available NewsItem layers.
     for schema in get_schema_manager(request).all():
+        # if filters and 'schema' in filters and filters['schema'].schema == schema:
+        #     visible = True
+        if no_layers_specified and show_default_layers and not show_custom_layer:
+            # default on if no 't' param given
+            visible = True
+        elif schemas and schema.id in schemas:
+            visible = True
+        else:
+            visible = False
         layers.append({
             'id': 't%d' % schema.id,
             'title':  schema.plural_name,
@@ -261,11 +294,57 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
                        'startdate': api_startdate,
                        'enddate': api_enddate},
             'bbox': False,
-            'visible': (no_layers_specified and show_default_layers) or schema.id in schemas # default on if no 't' param given
+            'visible': visible
         })
 
+    # Explicit filtering by ID.
+    ids = params.get('i') or u''
+    ids = [i.strip() for i in re.split(r'[^\d]+', ids)
+           if i.strip()]
+    if ids:
+        show_custom_layer = True
+        if filters is None:
+            filters = FilterChain(request)
+        filters.replace('id', *ids)
+
+
+    # 'Custom' layer. This is a catch-all for all filtering
+    # that isn't just enabling a default layer with the default
+    # date range.
+    # Not visible unless there is something like that to show.
+    if filters and sorted(filters.keys()) not in ([],
+                                                  ['date'],
+                                                  ['date', 'schema'],
+                                                  ['schema']):
+        show_custom_layer = True
+
+    if filters is not None:
+        # Don't inspect filters['schema']; that's already covered by schemas above.
+        base_url = reverse('map_items_json')
+        layer_url = filters.make_url(base_url=base_url)
+        # Quick ugly hacks to make the itemquery api happy.
+        # Hooray proliferation of spellings.
+        layer_url = layer_url.replace('locations=', 'locationid=')
+        layer_url = layer_url.replace('start_date=', 'startdate=')
+        layer_url = layer_url.replace('end_date=', 'enddate=')
+
+        if 'schema' in filters:
+            # Normally, filters.make_url() captures the schema in the
+            # path part of the URL. But map_items_json doesn't,
+            # so we add a query parameter.
+            params = {'type': [s.slug for s in filters['schema'].schemas]}
+        else:
+            params = {}
+        custom_layer = {
+            'url': layer_url,
+            'params': params,
+            'title': u"Custom Filter",
+            'visible': show_custom_layer,
+            'id': 'c1',
+            }
+        layers.append(custom_layer)
+
     is_widget = params.get('x', None) is not None
-        
     controls = {}
     control_list = params.get("v", None)
     if control_list is not None: 
@@ -306,11 +385,14 @@ def _decode_map_permalink(request, show_default_layers=True, filters=None):
       }, 
     }
 
-
+    if 'id' in filters:
+        # Put them in the params so the js code can construct, well,
+        # permalinks with these ids, on the client side.
+        ids = '-'.join(map(str, filters['id'].ids))
+        config['permalink_params']['i'] = ids
     if popup_info:
         config['popup'] = popup_info
 
-    
     if is_widget: 
         config['controls'] = controls
         if width is not None: 
@@ -418,7 +500,6 @@ def map_items_json(request):
     rendition of the REST api's geojson response. includes
     only attributes used by the map.
     """
-
     items, params = build_item_query(request)
 
     def _item_to_feature(item):