+++ ebpub/ebpub/urls.py
@@ -80,7 +80,13 @@ if get_metro()['multiple_cities']:
     urlpatterns += patterns(
         '',
         url(r'^streets/$', views.city_list, name='ebpub-city-list'),
-        url(r'^streets/([-a-z]{3,40})/$', views.street_list, name='ebpub-street-list'),
+        # optionally accept city slug below. this allows for a no arg
+        # ebpub-street-list call to return the street list for single city and
+        # city list for multi city setups. this way the templates can call
+        # {% url ebpub-street-list %} without having to worry about single/multi
+        # city configurations
+        url(r'^streets/(?:([-a-z]{3,40})/)?$', views.street_list,
+            name='ebpub-street-list'),
         url(r'^streets/([-a-z]{3,40})/([-a-z0-9]{1,64})/$', views.block_list,
             name='ebpub-block-list'),
         url(r'^streets/([-a-z]{3,40})/([-a-z0-9]{1,64})/%s/$' % BLOCK_URL_REGEX,
@@ -106,7 +112,7 @@ else:
     # single-city block patterns.
     urlpatterns += patterns(
         '',
-        url(r'^streets/(\w{0})$', views.street_list,
+        url(r'^streets/$', views.street_list,
             name='ebpub-street-list'),
         url(r'^streets/(\w{0})([-a-z0-9]{1,64})/$', views.block_list,
             name='ebpub-block-list'),